update v4 - firebase part 2 3/4 problems with databse

+++ app/src/main/java/com/example/project2/firestore/GameAttempt.kt
@@ -5,5 +5,5 @@ data class GameAttempt(
     val correctAnswers: Int? = null,
     val wrongAnswers: Int? = null,
     val totalAttempts: Int? = null,
-    val bestScore: Int? = null
+    val bestScore: Long? = null
 )
